Fix GCC 4.1 warning.
git-svn-id: http://root.cern.ch/svn/root/trunk@39437 27541ba8-7e3a-0410-8455-c3a389f83636

+++ core/textinput/src/textinput/History.cpp
@@ -74,7 +74,7 @@ namespace textinput {
       fHistFile.seekp(0);
       size_t nPrune = fPruneLength;
       if (nPrune == (size_t)kPruneLengthDefault) {
-        nPrune = fMaxDepth * 0.8;
+         nPrune = (size_t)(fMaxDepth * 0.8);
       }
       for (size_t Idx = fEntries.size() - nPrune,
         E = fEntries.size(); Idx < E; ++Idx) {
